(maint) Fix failing PuppetDB tests
When replacing facts in PuppetDB the request is non-blocking, and
so our acceptance tests will query for the updated facts for a target
before the command has actually be executed. This adds a retry to
the test to re-query for facts if they haven't been updated yet. The
retry tries 5 times, with a 5 second sleep, before failing. This should
be more than enough time for the command to be run.

This also updates the container image used for testing the
`run_container` plan function on *nix to use the `ubuntu:20.04` image
instead of `ubuntu:14.04`. Previously only the 14.04 tag was cached in
Github Actions environments, but as of April 12 that image has been
removed and `20.04` is now cached. This also updates the test itself to
verify that the image is available before executing tests to avoid
similar failures.

!no-release-note