(GH-2499) Add 'lookup' command
This adds a new `lookup` command to Bolt that looks up values with
Hiera. The command first runs `apply_prep` on each target to install the
puppet-agent package and collect facts, then sets up Hiera and looks for
the specified key in the hierarchy. The command requires a targeting
option, and the Hiera lookup is performed in the context of each target
passed to the command, allowing for interpolations using facts.

!feature

* **`lookup` command to look up values with Hiera**
  ([#2499](https://github.com/puppetlabs/bolt/issues/2499))

  The new `bolt lookup` and `Invoke-BoltLookup` commands can be used to
  look up values with Hiera.