(GH-2791) Allow Docker to connect with full IDs
Currently, Docker connections only succeed when the 12 character
"short ID" of a container is used. When a full container ID is
provided, connections fail despite the container being present
and running. This is because the Docker transport searches for the
container ID in the output of `docker ps` and `docker ps` only returns
short ID values unless the `--no-trunc` flag is used.

This commit updates the connection logic to pass the `--no-trunc` flag
and match container IDs using `start_with?` instead of an equality
check.

Fixes puppetlabs/bolt#2791.

!bug

* **Fix Docker connections using full ID strings**

  The Bolt Docker transport now successfully connects to containers
  when the full SHA 256 container ID string is provided as a name or
  URL.  Previously, Bolt could only connect when the 12 character
  shortened form of the ID string was used.