(GH-2785) Improve error message when provided git repo isn't available
When specifying modules to install, when a user provides a git module if
the module cannot be retrieved we raise an error. Previously the error
simply stated that the provided repo was invalid, without specifying
that private repositories cannot have their dependencies resolved and
must follow a documented workflow. This updates the error message to
state that the repo is either not a Git repo or is private, and link to
documentation describing how to install modules that include a private
Git repo without installing dependencies.

!no-release-note