(maint) Use Net::SSH pageant module to check pageant
Previously we were manually using Window's
[FindWindowW](https://docs.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-findwindoww)
function via Fiddle to see if pageant is running and determine if we
need to pass `use-agent: false` to Net::SSH on Windows. Initially this
check ran just fine in CI, but recently started exiting 1 when called.
Our theory is that the types we had for the parameters or return were
incorrect, but since the system was exiting it was impossible to catch
the error and see it, and since this couldn't be reproduced locally and
was only happening in Github Actions finding system logs was also not
possible.

This updates our own pageant check to simply use the
Net::SSH::Authentication::Pageant::Win module which defines pretty much
the same function (`FindWindow`) and use that to see if pageant is
running.

!no-release-note