(hack) Bash tab completion for Bolt commands
This updates our tab completion script to be accurate with modern Bolt
commands and flags. It updates the script to only produce the options
available for a particular Bolt subcommand, and to only include the
flags after the entire Bolt command has been entered. This also adds
support for tab completing Bolt tasks and plans when they are cached in
the users current directory or the default Bolt project. This means Bolt
can't tab complete tasks and plans if the user is using a Boltdir, or if
they haven't run a command that caches tasks and plans yet.

In order to support tab completing task names, this also adds caching
for Task data that behaves exactly like Plan data caching. The cache is
registered or updated whenever types are generated, or when `bolt task
show` has been run without a cache. Bolt will check the mtime of all
task files to verify that the cache is up to date, and use the cache
data if available.

!feature

* **Builtin tab completion**

  Bolt now supports tab completion for Bolt commands, subcommands, and
  any flags available for a command. It also supports tab completing tasks
  and plans for certain project types when tasks and plans have been
  cached.