(hack) Add workflow to publish PuppetBolt PowerShell module
This adds a new workflow to release the PuppetBolt PowerShell module.
The module is manually triggered and defaults to building the module
from the `main` branch, but has an optional `ref` parameter that can be
used to specify a specific ref (such as a tag).

When attempting to publish a version of the module that has already been
published, this workflow will fail. The published module is not
overwritten.

!no-release-note