(bug-fix) Support passing '_run_as' to 'apply_prep' plan function
Previously, Bolt would not pass the `_run_as` option through to the
executor when the option was supplied to `apply_prep()`. The `run_as`
configuration would get successfully picked up through other methods,
like passing on the CLI or inventory, or included in the
`puppet_library` plugin configuration. This merges any supported options
passed to `apply_prep` with options passed to the `puppet_library`
plugin hook, giving the plan function configuration precedence. For now
the only supported metaparameter is `_run_as` since that is the only
option supported by the `puppet_library` plugin.

!bug

* **Support _run_as passed to apply_prep()**
  Bolt now respects the `_run_as` metaparameter when passed to the
  `apply_prep()` plan function. This is the only supported metaparameter,
  and takes highest precedence per the [Bolt configuration
  precedence](https://puppet.com/docs/bolt/latest/configuring_bolt.html#configuration-precedence)