(bug-fix) Don't stacktrace if welcome message file can't be written
In order to determine whether to print the Bolt welcome message,
Bolt inspects a directory for a file, and if the file does not exist
prints the welcome message and lays down the file so that future runs
don't also print the welcome message. This allows the welcome message to
be printed on a mostly-per-user basis, and avoids environment variable
instability.

Previously the path for the file was based on `Bolt::Config.user_path`,
but when that fails to be created Bolt will stacktrace. Now, Bolt
catches `StandardError` errors and falls back first to
`Bolt::Config.system_path`, and then to returning `nil` and avoiding
printing the welcome message entirely. This ensures that trying to print
a welcome message doesn't result in Bolt being unusable due to erroring
out trying to create a file in an invalid location.

!bug

* **Don't stacktrace if welcome message file can't be written**

  Previously, Bolt would stacktrace if it failed to make the directory
  to store the welcome message file in, which relies on expanding `~`.
  Bolt now falls back to a system-level path, and then omits the welcome
  message entirely if the system-level path also fails to be created or
  written to.