(bug) Do not lookup `:future` in Puppet for `file::*` functions
This fixes a bug in the `file::*` plan functions that was introduced
in #2731. Previously, these functions would look for future
configuration with a Puppet lookup (i.e. `Puppet.lookup(:future)`).
However, Bolt does not set this value in Puppet, causing the functions
to error. This updates the functions to only check the Bolt executor for
future configuration.

!bug

* **Do not error when looking up future configuration in `file::*` plan
  functions**

  The `file::read`, `file::readable`, and `file::exists` plan functions
  no longer error when checking for `future` configuration in Bolt.