(GH-2820) Update link to module spec documentation
This updates the link to the documentation for module specs in the
project configuration file. Previously, if a project included an invalid
spec, the raised error message would direct the user to the 'Modules
overview' page. Now, the message directs the user to the docs that
explicitly define valid specs.

!no-release-note