(maint) Update illegal characters message to list legal characters
When provided a filter that includes characters that aren't allowed in
task or plan names, Bolt raises an error stating that the provided
filter includes characters that cannot be used in task or plan names.
This updates the message to include a list of valid characters in the
message so that users can better understand why the error was raised.

!no-release-note