(GH-2806) Support run_as for container transports using Bash
This adds support for using `run-as` and associated configuration
options to the Docker, Podman, and LXD transports when running on *nix
systems. Originally, 'run-as' was not supported for the Docker transport
because it could be running on either Windows or *nix, and run-as is not
supported when running on Windows systems. At the time it seemed
confusing for the option to only be supported sometimes, but now that
this is the case for the SSH transport when using the PowerShell
`login-shell` there's more of a precedent for users understanding when
it is and isn't supported. This was intended to be
addressed after the great [shell-transport split of 2020](https://github.com/puppetlabs/bolt/pull/1702),
and just fell through the cracks.

In addition to adding support for `run-as` to Docker when running on
*nix, this adds support for the option on the related LXD and Podman
transports under the same circumstances, since those transports are
largely based on the Docker transport and operate in the same way. LXD
specifically does not support PowerShell, so `run-as` support requires
no caveats there.

This also fixes a bug where Bolt would stacktrace if `run-as` was set
for any of the above transports, because the Bash shell would try to
call `reset_cwd?` on the transport which was not defined.

!feature

* **Support run-as for container transports when running on \*nix** ([#2806](https://github.com/puppetlabs/bolt/issues/2806))
  The Docker, LXD, and Podman transports now support `run-as`
  configuration and related configuration options when running on *nix
  systems. `run-as` is not supported for any Windows systems or the
  PowerShell shell over SSH.