(PE-31709, PE-31710) Standardize bolt response errors
This commit updates bolt-server error responses to follow a newly defined pattern for detailing issues. In general error responses are now categorized by status code. Errors with a 400 status are ones that indicate a problem with the request: for example schema errors or missing/wrong query params. Errors with execution or general configuration problems with the service itself are represented with a 500. The implementation has been refactored to utilize error handlers at the route level that return the appropriate error code based on the class of exception raised. A new error type BoltServer::ResponseError has been introduced which, when raised will result in a 400 error response.

Some other notable updates around standardization:
 - When asking for information about plans/tasks in bolt-server `environment` is required. Previously an unstructured string response which does not match other serialized error objects was returned. This commit updates that error to conform to the expected structure.
 - Similarly the use of backticks in messages is generally undesireable, backticks have been removed and replaced with single quotes where applicable.