(GH-2832) Use module name for plugin tar directory
This updates the applicator to use the name of a module as the name of
the directory that plugin files are saved in when building the plugin
tarball. Previously, the applicator would use name of the module
directory as name of the directory that plugin files were saved to. This
caused issues when referencing a file in a project that had a configured
name that is different than the project directory's name.

For example, in a project named `foo` that uses a `Boltdir`, a file
might be located at `foo/Boltdir/files/somefile.txt`. When building the
plugin tarball, the applicator would set the path of this file to
`Boltdir/files/somefile.txt` instead of the correct path
`foo/files/somefile.txt`. This would cause any plan that referenced that
file using the correct syntax `puppet:///modules/foo/somefile.txt` to
error.

!feature

* **Upload project plugin files to correct directory when running an
  apply**
  ([#2832](https://github.com/puppetlabs/bolt/issues/2832))

  Project plugin files are now uploaded to the correct directory when
  running an apply. Previously, if a project used a `Boltdir` or had a
  directory name that did not match the project's configured name, apply
  blocks could not correctly reference files in the project using Puppet
  file syntax (`puppet:///modules/<project name>/<file name>`).