(bug) Use Pathname object when building plugin tar ball
This fixes a bug in `Bolt::Applicator.build_plugin_tarball` that was
causing Bolt to stacktrace when running an apply block on a Windows
controller. When Bolt builds the tar ball, it joins the module's name
with the path to the module file relative to the module directory (i.e.
`my_module` + `files/some_file.txt`). To get the relative path to the
module file, Bolt calls the Ruby function `Pathname.relative_path_from`.
Previously, the string representation of the file's full path was passed
to this function, which then errored as it called `Pathname.cleanpath`
internally, which is not valid for string objects. Bolt now passes a
`Pathname` object to this function, fixing the stacktrace issue.

!no-release-note