(GH-2839) Move logic from `Bolt::CLI` to new `Bolt::Application` class
This moves a bulk of the logic for executing Bolt's commands from the
`Bolt::CLI` class to a new `Bolt::Application` class. Previously, the
`Bolt::CLI` class had multiple responsibilities:

- Parsing the command
- Validating the command
- Initializing all of Bolt's classes
- Determing which command to execute
- Executing each command, including much of the logic to do so

This left the `Bolt::CLI` class overloaded, as it was essentially the
sole interface for running Bolt. This made it difficult at times to
modify existing behavior or add new commands, as the class was
incredibly tangled. Classes needed to be loaded and methods needed to be
invoked in a very particular order, otherwise Bolt would break.

This marks the first step in a larger refactoring of the `Bolt::CLI`
class. Further refactoring might include:

- Move responsibilities for validating commands to the parser
- Split the `Bolt::Application` class into several smaller subclasses
  that are each responsible for running a specific subcommand
- Offload resposibility for loading some classes, such as `Bolt::Rerun`
  and `Bolt::PAL`, to `Bolt::Application`
- Defering resolving inventory until targets are needed

!no-release-note