(GH-2866) Include undef variables in parallelize blocks
When creating a new Future block, Bolt must copy any variables that
should be in scope for the block into a new scope for the block, so that
if variables redefined inside the block they don't affect other blocks
running in parallel. Gathering variables in scope uses the
`Puppet::Parser::Scope#to_hash` method, which accepts 2 positional
parameters:
- The first parameter tells the function whether to recursively get
  variables from parent (aka enclosing) scopes. This defaults to true
  and we want it to be true.
- The second parameter tells the function whether to include variables
  set to 'undef', which defaults to false and which we want to be true.

Previously we had not enabled including undef variables in the new
scope, which resulted in defined variables not being passed to the new
block as users expect. This changes how we call `to_hash` to correctly
include variables that are `undef`.

!bug

* **Undef variables are now included in Future block scopes** ([#2866](https://github.com/puppetlabs/bolt/issues/2866))
  Previously, Bolt would not include variables with a value `undef` in
  Future block scopes, leading them to be undefined. Bolt now includes
  those variables when creating the new scope for Future blocks.