(bug) Do not error when using --run-as on Windows controller
This fixes a bug introduced by #2806 that causes Bolt to error on a
Windows controller whenever the `--run-as` command-line option is used.
Previously, the `docker` and `podman` transport config classes included
validation that would raise an error whenever the `run-as` option was
set for the transport and the controller was running Windows. Because
Bolt automatically sets the `run-as` configuration for all transports
that support it when using the `--run-as` command-line option, this
resulted in any command using that option to raise this error, even if
there weren't any targets using that transport.

This removes the relevant validation in the transport config classes.
This will not result in any errors when running on targets using these
transports, as Bolt automatically sets the transport's shell to
`powershell`, which does not include logic for running as another user
and cannot be overridden by the user.

!bug

* **Do not error when using `--run-as` on a Windows controller**

  Bolt no longer raises an error when using the `--run-as` command-line
  option on a Windows controller.