(GH-2877) `wait()` with no arguments waits on all Futures from the plan
This amends the `wait()` plan function to wait on all the Futures that
were spawned from the same plan as the `wait()` call if the function is
not provided a list of futures. The function still accepts a `timeout`
or `options` parameter if it is not passed any futures. To support this,
Bolt's `PlanFuture` object now has a `plan_id` attribute, which
determines which **invocation** of a plan the Future was created from
(since the same plan can be invoked multiple times in parallel). When
`wait()` is called without Futures, it selects the Futures from the
FiberExecutor with a matching `plan_id` to itself, and waits on those.

This also removes a flaky test around whether the cache file is updated
when plans in modules are modified. The test doesn't seem to be failing
legitimately based on local testing.

- [ ] puppetlabs/orchestrator#1738
- [ ] puppetlabs/pe-bolt-modules#20

!feature

* **`wait()` without arguments waits on all Futures from plan** ([#2877](https://github.com/puppetlabs/bolt/issues/2877))

  Calling the `wait()` plan function without a list of Futures will now
  wait on all Futures created so far in the plan.