(GH-2899) Add `Bolt::Util::Format` module for formatting utilities
This adds a new `Bolt::Util::Format` module that includes utility
methods for formatting. Specifically, this moves the `stringify()`
method that used to live on the `out::message` function to a reusable
module, as the `log::*` functions need this functionality as well.

Note: These methods cannot live on `Bolt::Outputter`, as they must be
able to format Puppet object types in plans. Attempting to stringify a
Puppet object type outside the scope of a plan results in a "stack trace
too deep" error.

!no-release-note