(bug) Fix stacktrace when cache includes missing file
This fixes a bug where Bolt would stacktrace when listing plans or tasks
and the plugin or task cache included a file that no longer exists.
Specifically, this error is raised by the `File.mtime` method receiving
a non-existent filepath.

The conditions that cause this bug are:

- Moving a module to a different location in the modulepath.
- Deleting a module that shadows another module of the same name on the
  modulepath.

If the user then runs `plan|task show`, Bolt will error.

!bug

* **Do not stacktrace with non-existent files in plan or task cache**

  Bolt no longer errors with a stacktrace when listing plans or tasks
  when the plan or task cache include paths to files that no longer
  exist.