(PE-32176) Use versioned_project from body for /project_inventory_targets
At one time, this endpoint used a GET request and therefore expected
versioned_project as a query param. It was later changed to a POST
request with puppet_connect_data included in the request body, but the
versioned_project query param remained. For consistency, we now expect
both pieces of data to be supplied in the request body.

!no-release-note