(maint) Fix broken Windows acceptance tests
Starting with Bolt 3.0, we no longer include the `bolt` function in the
PuppetBolt PowerShell module and instead shipped with a batch file for
Windows packages that would run the Bolt executable. When this change
was made acceptance tests were broken as the `PATH` did not include the
path to Bolt's `bin/` directory and the test would error with a "bolt is
not a recognized command" message.

During the beaker setup process (specifically step 032), beaker
updates the `PATH` to include Bolt's `bin/` directory. Previously, the
script responsible for this used the `Join-Path` cmdlet with multiple
positional arguments, which is not supported in PowerShell < 6.0,
resulting in the `PATH` not getting updated. Updating this line to not
error correctly sets the `PATH` for the step that sets it, but the
`PATH` is not preserved between PowerShell sessions. This resulted in
the tests continuing to fail since `bolt` was not recognized as a
command.

This change updates the `bolt_command_on` helper method to set the
`PATH` prior to invoking `bolt`, ensuring that the `PATH` is correctly
configured every time Bolt is run.

!no-release-note