(GH-2875) Add `batch-mode` config option for `native-ssh`
This adds a new `batch-mode` configuration option for the `native-ssh`
transport. When set to `false`, `BatchMode` is disabled, which allows
SSH to fall back to querying for a password when key authentication
fails. This config option can be set in conjunction with the
`ssh-command` option to allow users to provide a password
non-interactively to SSH when using `native-ssh`.

!feature

* **Add `batch-mode` configuration option for `native-ssh` transport**
  ([#2875](https://github.com/puppetlabs/bolt/issues/2875))

  The `native-ssh` transport has a new `batch-mode` configuration option
  that can be used to enable or disable `BatchMode`. For more
  information, see [the
  documentation](https://puppet.com/docs/bolt/latest/troubleshooting.html#providing-a-password-non-interactively-using-native-ssh).