(PE-32485) Allow plugin_hooks to be passed to bolt-server actions
This adds the plugin_hooks key to the SSH and WinRM target schemas,
allowing it to be passed to the API. This will let the key be used
internally within orchestrator without worrying about whether it gets
passed to bolt-server.

This key, along with the existing special 'hostname' and
'private-key-content' keys, will automatically be ignored by Bolt's
transport config handling. However, for clarity we now explicitly filter
the options we pass to the config to include only those that are
accepted by the transport. While not strictly necessary, this makes it
more obvious why those keys don't cause a problem and it provides an
extra layer of certainty in case the underlying implementation shifts in
the future.