(hack) Format PlanResult if it's a Bolt datatype
Bolt will format a PlanResult as JSON when we don't have a handler for
the class of the result. Despite having functions for printing
`Bolt::Result`, `Bolt::ApplyResult`, and `Bolt::Error` types, the plan
result printer in the human outputter would not use these functions and
would format the types as JSON, which often makes them difficult to
read. Bolt will now correctly format the result types if returned from a
plan.

!feature

* **Format PlanResults if they are Bolt datatypes** (hack)

  Bolt will now print a more human readable message for plan results
  that are Bolt datatypes when using the human output format, rather than
  printing JSON.