(GH-2889) Temporarily override scope when resuming future
This updates `PlanFuture` instances to keep track of their own scope
object. When a `PlanFuture` is resumed by the `FiberExecutor`, the
`global_scope` in Puppet is temporarily overridden with the `PlanFuture`
scope. This prevents ephemerals from being popped off in the wrong order
due to race conditions that arise from invoking custom Puppet language
functions that invoke Bolt executor functions.

!bug

* **Do not remove variables from scope when using `parallelize()`**
  ([#2889](https://github.com/puppetlabs/bolt/issues/2889))

  Bolt no longer removes variables from a plan's scope when invoking
  a custom Puppet language function that invokes a `run_*` function from
  a `parallelize()` block. Previously, a race condition would result in
  the current plan's variables to be removed from the plan's scope.