(GH-2925) Support `_catch_errors` in `apply_prep`
This adds support for using the `_catch_errors` metaparameter in the
`apply_prep` plan function. When set, the function will return a result
set that includes results for all targets that successfully installed
targets and retrieved custom facts, and any errors for targets that were
not able to do both.

This also updates the `bolt apply` and `Invoke-BoltApply` commands to
set `_catch_errors` when running the `apply_prep` function. This allows
the command to complete the apply on as many targets as possible instead
of erroring if even one target fails the `apply_prep` step.

!feature

* **Support `_catch_errors` in `apply_prep`**
  ([#2925](https://github.com/puppetlabs/bolt/issues/2925))

  The `apply_prep` function now accepts the `_catch_errors` parameter.
  Additionally, the `bolt apply` and `Invoke-BoltApply` commands set
  `_catch_errors` to `true`, allowing an apply action to complete on as
  many targets as possible instead of erroring if even one target fails
  the `apply_prep` portion of the command.