(GH-2945) Respect `@api` strings tag, deprecate `@private` tag
This adds support for using the `@api private|false` tag in plans and
deprecates the `@private false|true` tag. Originally, plans were marked
as private using the `@private` tag, but this created differing behavior
from what is established and documented by puppet-strings.

When a plan uses the `@private` tag, Bolt will now issue a deprecation
warning, and suggest the user use the `@api` tag instead.

If both tags are present, Bolt will consider the plan private if either
`@api private` or `@private true` are present. This is because
puppet-strings sets the `@api` tag to `public` by default, whether it's
actually present in the plan or not.

!feature

* **Respect `@api` tag in Puppet language plans**
  ([#2945](https://github.com/puppetlabs/bolt/issues/2945))

  Bolt now respects the `@api` tag in Puppet language plans. If the
  `@api private` tag is present, the plan will be hidden from `bolt plan
  show` and `Get-BoltPlan` output.

!deprecation

* **Deprecate `@private` tag in Puppet language plans**
  ([#2945](https://github.com/puppetlabs/bolt/issues/2945))

  The `@private` tag for Puppet language plans is deprecated. Use the
  `@api` tag instead.