(hack) Add --script to plan new
This adds a new flag `--script` to the `bolt plan new` command that
accepts a detailed (new-style) Puppet file reference to a script. The
flag can only be used if `future.file_paths` is configured, so that
detailed file references are loaded. Bolt will then validate the file
exists, and create a new plan that simply runs the script and returns
the result.

Many users migrate to using Bolt from running scripts, and find it
difficult to begin writing Bolt content that can orchestrate their
scripts. This makes it easy to create the boilerplate for running a
script, with future improvements to accept parameters for the script and
integrate those into the generated plan.

!feature

* **Pass `--script` to `bolt plan new` to wrap a script in a plan** (hack)
  The `bolt plan new` command now accepts a `--script SCRIPT` parameter
  that will generate a plan that simply runs the script and returns the
  result. The flag only supports detailed (new-style) Puppet file
  references. Learn more about file references at http://pup.pt/bolt-scripts