(maint) Restart a task over PCP if it fails with 'plan-already-finished'
When running a plan over the PCP transport, Bolt issues a
`command/plan_start` API request and gets a Plan ID to use throughout
the Bolt run. The plan ID is then associated with every subsequent
`run_task` request, so that the task is associated with the plan it's
running as part of. If the Orchestrator is restarted as part of the
plan, it (correctly) marks the plan as 'failed' and returns a
`plan-already-finished` error for subsequent task runs using the same
ID. This makes sense, but Bolt can continue to run the tasks with a new
ID given that restarting the Orchestrator is a somewhat common case.
When Bolt gets a `plan-already-finished` error from the Orchestrator it
will start a new plan to generate a new Plan ID, and will rerun the task
and continue executing.

!feature

* **Continue executing when a plan over PCP restarts the Orchestrator**

  Bolt will now retry a task once when running over PCP as part of a
  plan, if the task errors with `plan-already-finished`.