(PE-32492, PE32496) Generalize project cache, fix concurrent writes
This commit udpates the file cache methods to generally cache based on a scheme of unique project_ref and file name. This is currently used for plugin tarballs and task metadata. Importantly this wraps the file write for caching project files in a serial_execute to avoid multiple threads trying to write a file.