(maint) Remove broken bundled-ruby acceptance test
This removes a broken acceptance test that checks that Ruby environment
variables are restored when the `bundled-ruby` is disabled.

This test currently fails for two reasons:

- The `gem` executable is not available on all *nix VMs when running
  tests on Jenkins. This causes the task to fail when attempting to set
  `GEM_PATH` with `gem env GEM_PATH`. This issue can be solved by
  setting `GEM_PATH` to a dummy value when running the command on the
  SUT.

- The `bundled-ruby` config is not disabled. This causes the test to run
  using Bolt's bundled Ruby, which is not what is being tested. This
  issue can be solved by setting `bundled-ruby: false` in the SUT's
  inventory file. However, doing so breaks other `local` transport tasks
  that rely on Bolt's bundled Ruby.

Since this feature has been reliably manually tested, and has a
sufficient integration test, we should remove this broken acceptance
test until we can find a suitable alternative.

!no-release-note