(bug) Do not set `env_vars` to an empty hash when none are provided
This fixes a bug introduced in the CLI refactor. Previously, if a user
did not provide any env vars to a command or script, Bolt would not set
the `env_vars` option for the command. When the CLI was refactored and
the `Bolt::Application` class was introduced, this option was set to
have a default value of an empty hash `{}`. This resulted in the
`Bolt::Shell::Bash` class wrapping all commands in `sh -c`, which caused
commands to fail on some targets (such as those using `native-ssh` to
connect to Windows targets that did not recognize the `sh` command).

!bug

* **Do not wrap commands in `sh -c` when not using environment variables
  or privilege escalation**

  Bolt no longer wraps commands in `sh -c` when environment variables
  are not specified and not using privilege escalation. Previously, Bolt
  would wrap all commands in `sh -c` when using the SSH transport, which
  caused commands to fail on targets that do not recognize `sh`, such as
  some Windows targets.