(GH-2992) Defer resolving plugin config and plugin hooks
This updates the Bolt to only resolve plugin config (`plugins`) and
plugin hooks (`plugin-hooks`) as needed. Previously, when Bolt setup the
`Bolt::Plugin` class, it resolved all plugin config and hooks as part of
the setup process. Because Bolt now loads the `Bolt::Plugin` class for
all commands, this resulted in Bolt erroring in cases where a plugin
module was not available, even if the command run didn't need access to
any plugins. This is particularly problematic for the `module install`
command, as users installing the project for the first time would be
attempting to install the very plugin modules that Bolt will say are
missing when the command is run.

!bug

* **Resolve plugins as needed**
  ([#2992](https://github.com/puppetlabs/bolt/issues/2992))

  Bolt now only resolves plugin configuration, plugin hooks, and
  inventory plugins as needed. Previously, Bolt resolved all plugins
  each time a command was run, causing some commands to fail when
  they should not.