(GH-2992) Defer resolving inventory plugins
This updates the inventory to defer resolving plugins until the
inventory is needed. During the CLI refactor, Bolt was changed to always
load the inventory, regardless of whether the inventory was actually
used. This resulted any plugins in an inventory being resolved whenever
a Bolt command was run, potentially leading to a significant degradation
in performance.

Now, when the inventory is instantiated, Bolt only stores the raw
inventory data. Once Bolt requests access to any of the groups, targets,
or config in the inventory, the inventory parses the data, resolves
plugin references, and loads the groups and targets into inventory.

!no-release-note